Better cycle prevention and counting mechanism implemented
